Allow the usage of the action in a reusable workflow
In a reusable workflow, the github context is the one of the caller,
which means that the gitub.repo is the caller repository that can differ
to the one where the reusable workflow is located. In this case we can
now defined a new input giving the workflow repository location.